+++ cms/src/layouts/ProtectedLayout.tsx
@@ -22,7 +22,7 @@ const ProtectedLayout: React.FC<{ children?: React.ReactNode }> = ({
   const { accessToken, isAuthenticated, setAccessToken } = useAuth();
   const { user, isUserDataLoading } = useUserData();
   const location = useLocation();
-  const axiosPrivate = useAxiosPrivate();
+  const axiosPrivate = useAxiosPrivate(accessToken, setAccessToken);
   const throwError = useError();
   // console.log(prevLocation);
   // if (!accessToken)