+++ cms/src/routes/_protected.tsx
@@ -1,7 +1,7 @@
 import { createFileRoute, Outlet, redirect } from "@tanstack/react-router";
 
 export const Route = createFileRoute("/_protected")({
-  beforeLoad: ({ context, location }) => {
+  beforeLoad: async ({ context, location }) => {
     console.group("/dashboard beforeLoad running...");
     if (!context.auth.isAuthenticated) {
       console.log(
@@ -18,7 +18,7 @@ export const Route = createFileRoute("/_protected")({
     }
 
     // const result = false;
-    // const result = context.auth.authorize();
+    // const result = await context.auth.authorize();
     // console.log("result:", result);
     console.groupEnd();
   },